backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
prisma==0.15.0
pydantic==2.11.1
python-dotenv==1.1.0
